# storage/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LouvorJA API",
        "description": "API do LouvorJA - Gerenciador de Hinários e Músicas para a Igreja Adventista do Sétimo Dia. Fornece endpoints públicos para consulta de músicas, letras, arquivos e álbuns, além de endpoints administrativos protegidos por autenticação JWT.",
        "contact": {
            "name": "LouvorJA",
            "email": "contato@louvorja.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.louvorja.com.br",
            "description": "Produção"
        },
        {
            "url": "http://localhost:8000",
            "description": "Desenvolvimento"
        }
    ],
    "paths": {
...
        "/db/manifest": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Manifest de arquivos JSON",
                "description": "Lista todos os arquivos JSON disponíveis em public/db/json com nome, tabela e path de acesso.",
                "operationId": "3cf90931a697db0625f6a69acbb1f360",
                "responses": {
                    "200": {
                        "description": "Lista de arquivos JSON disponíveis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "file": {
                                                "type": "string",
                                                "example": "config.json"
                                            },
                                            "table": {
                                                "type": "string",
                                                "example": "config"
                                            },
                                            "path": {
                                                "type": "string",
                                                "example": "/db/config"
                                            },
                                            "hash": {
                                                "description": "MD5 hash do arquivo para cache ETag",
                                                "type": "string",
                                                "example": "a1b2c3d4e5f6..."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
        "/db/{table}": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Registros de uma tabela JSON",
                "description": "Retorna registros paginados de um arquivo JSON específico. Suporta paginação via query params.",
                "operationId": "d1af8531b38e332f3446bf819af8fcd7",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Nome da tabela (arquivo JSON sem extensão)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registros paginados",
                        "headers": {
                            "ETag": {
                                "description": "MD5 hash do arquivo para cache condicional",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Não modificado (ETag coincide com If-None-Match)"
                    },
                    "404": {
                        "description": "Arquivo não encontrado"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
        "/db/bundle": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Bundle ZIP de todos os JSON",
                "description": "Gera e baixa um arquivo ZIP contendo todos os arquivos JSON da pasta public/db/json.",
                "operationId": "35c225703d4ac77f22ab32fe1de74730",
                "responses": {
                    "200": {
                        "description": "Arquivo ZIP",
                        "content": {
                            "application/zip": {}
                        }
                    },
                    "404": {
                        "description": "Nenhum arquivo JSON encontrado"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
        "/db/{table}/categories": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Categorias únicas de uma coluna",
                "description": "Retorna array de valores únicos de uma coluna específica em um arquivo JSON, ordenados alfabeticamente.",
                "operationId": "6600cca444ceda9e5692d30654b70ff8",
                "parameters": [
                    {
                        "name": "table",
                        "in": "path",
                        "description": "Nome da tabela",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "column",
                        "in": "query",
                        "description": "Nome da coluna para extrair categorias únicas",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array de categorias únicas ordenadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parâmetro column obrigatório"
                    },
                    "404": {
                        "description": "Tabela não encontrada"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
        "/json_db/{file}": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Obter arquivo JSON exportado (legado)",
                "description": "Retorna o conteúdo de um arquivo JSON específico da pasta public/db/json/",
                "operationId": "5c607d2e3a8b8ce46990e2a52ed4788f",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "Nome do arquivo (sem extensão .json)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conteúdo do JSON"
                    },
                    "404": {
                        "description": "Arquivo não encontrado"
                    }
                },
                "security": []
            }
        },
        "/db/export": {
            "get": {
                "tags": [
                    "Database"
                ],
                "summary": "Recriar JSONs do banco",
                "description": "Recria todos os arquivos JSON estáticos gerados a partir do banco de dados.",
                "operationId": "d24f92e46c4d3b46b57800babe555bab",
                "responses": {
                    "200": {
                        "description": "JSONs recriados com sucesso"
                    },
                    "500": {
                        "description": "Erro ao recriar arquivos"
                    }
                },
                "security": [
                    {
                        "ApiToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação JWT via header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}